examples/ap2-standard-payment/pyproject.toml
[project]
name = "ap2-standard-payment"
version = "0.1.0"
description = "Standard AP2 payment implementation over A2A/HTTP - baseline comparison for P2P version"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "starlette>=0.37",
    "uvicorn>=0.29",
    "PyJWT>=2.8",
]

[tool.setuptools.packages.find]
include = ["ap2_types*", "a2a_helpers*", "agents*"]
